+++ maslevtsov.stanislav/T3/commands.cpp
@@ -1,82 +1,9 @@
 #include "commands.hpp"
 #include <algorithm>
-#include <functional>
 #include <numeric>
+#include <functional>
 #include <io_fmt_guard.hpp>
-
-namespace {
-
-  bool is_even_vertex_num(const maslevtsov::Polygon& polygon)
-  {
-    return polygon.points.size() % 2 == 0;
-  }
-
-  bool is_odd_vertex_num(const maslevtsov::Polygon& polygon)
-  {
-    return !is_even_vertex_num(polygon);
-  }
-
-  bool is_equal_vertex_num(std::size_t vertex_num, const maslevtsov::Polygon& polygon)
-  {
-    return vertex_num == polygon.points.size();
-  }
-
-  double add_main_diagonal(const maslevtsov::Point& lhs, const maslevtsov::Point& rhs)
-  {
-    return lhs.x * rhs.y;
-  }
-
-  double subtract_sub_diagonal(const maslevtsov::Point& lhs, const maslevtsov::Point& rhs)
-  {
-    return -lhs.y * rhs.x;
-  }
-
-  double get_polygon_area(const maslevtsov::Polygon& polygon)
-  {
-    std::vector< maslevtsov::Point > second_lace;
-    std::copy(++polygon.points.cbegin(), polygon.points.cend(), std::back_inserter(second_lace));
-    second_lace.push_back(*polygon.points.cbegin());
-    std::vector< double > to_accumulate(polygon.points.size());
-    auto begin_it = polygon.points.cbegin();
-    auto end_it = polygon.points.cend();
-    double area = 0.0;
-    std::transform(begin_it, end_it, second_lace.cbegin(), to_accumulate.begin(), add_main_diagonal);
-    area += std::accumulate(to_accumulate.cbegin(), to_accumulate.cend(), 0.0);
-    std::transform(begin_it, end_it, second_lace.cbegin(), to_accumulate.begin(), subtract_sub_diagonal);
-    area += std::accumulate(to_accumulate.cbegin(), to_accumulate.cend(), 0.0);
-    area *= 0.5;
-    area = std::abs(area);
-    return area;
-  }
-
-  bool compare_vertex_num_less(const maslevtsov::Polygon& lhs, const maslevtsov::Polygon& rhs)
-  {
-    return lhs.points.size() < rhs.points.size();
-  }
-
-  bool is_same(const maslevtsov::Polygon& lhs, const maslevtsov::Polygon& rhs)
-  {
-    if (lhs.points.size() != rhs.points.size()) {
-      return false;
-    }
-    return std::mismatch(lhs.points.cbegin(), lhs.points.cend(), rhs.points.cbegin()).first == lhs.points.cend();
-  }
-
-  struct WithSamesInserter
-  {
-    std::vector< maslevtsov::Polygon >& with_echos;
-
-    int operator()(maslevtsov::Polygon& polygon, const maslevtsov::Polygon& to_compare)
-    {
-      auto same_with_arg = std::bind(is_same, to_compare, std::placeholders::_1);
-      with_echos.push_back(polygon);
-      if (same_with_arg(polygon)) {
-        with_echos.push_back(polygon);
-      }
-      return 0;
-    }
-  };
-}
+#include "functors.hpp"
 
 void maslevtsov::get_area(const std::vector< Polygon >& polygons, std::istream& in, std::ostream& out)
 {
